refactor: update core imports to use tracecat.interactions
- Change imports from tracecat.ee.interactions to tracecat.interactions
- Update db/schemas.py, dsl/, expressions/, validation/, workflow/
- Use core types for InteractionContext, ActionInteraction, etc.
- Reduces import-time coupling between core and EE
